+++ packages/react-wrappers/src/utils.ts
@@ -107,24 +107,6 @@ export function saveReactUtils(outdir: string) {
   const reactUtils = `
 import { useEffect, useLayoutEffect } from "react";
 
-export function useAttribute(targetElement, attrName, value) {
-  useLayoutEffect(() => {
-    if (value !== undefined && attrName !== 'style' && targetElement.current?.getAttribute(attrName) !== String(value)) {
-      targetElement.current?.setAttribute(attrName, String(value));
-    }
-  }, [value]);
-}
-
-export function useBooleanAttribute(targetElement, attrName, propName) {
-  useEffect(() => {
-    if (!propName || propName === 'false') {
-      targetElement.current?.removeAttribute(attrName);
-    } else {
-      targetElement.current?.setAttribute(attrName, '');
-    }
-  }, [propName]);
-}
-
 export function useProperties(targetElement, propName, value) {
   useEffect(() => {
     if (value !== undefined && targetElement.current[propName] !== value) {